Rename client_rework to ngclient
The current client and the next-gen client should
coexist in the same repository during the ongoing
development of the latter.

Looking for a name which is client-related,
short, meeting PEP8 package names requirements.
Currently "ngclient" seems to fit in until a
better proposal comes.

Rename updater_rework.py to updater.py
