+++ tox.ini
@@ -43,14 +43,14 @@ commands =
     # Use different configs for new (tuf/api/*) and legacy code
     # TODO: configure black and isort args in pyproject.toml (see #1161)
     black --check --diff --line-length 80 {toxinidir}/tuf/api
-    black --check --diff --line-length 80 {toxinidir}/tuf/client_rework
+    black --check --diff --line-length 80 {toxinidir}/tuf/ngclient
     isort --check --diff --line-length 80 --profile black -p tuf {toxinidir}/tuf/api
-    isort --check --diff --line-length 80 --profile black -p tuf {toxinidir}/tuf/client_rework
+    isort --check --diff --line-length 80 --profile black -p tuf {toxinidir}/tuf/ngclient
     pylint {toxinidir}/tuf/api --rcfile={toxinidir}/tuf/api/pylintrc
-    pylint {toxinidir}/tuf/client_rework --rcfile={toxinidir}/tuf/api/pylintrc
+    pylint {toxinidir}/tuf/ngclient --rcfile={toxinidir}/tuf/api/pylintrc
 
     # NOTE: Contrary to what the pylint docs suggest, ignoring full paths does
     # work, unfortunately each subdirectory has to be ignored explicitly.
-    pylint {toxinidir}/tuf --ignore={toxinidir}/tuf/api,{toxinidir}/tuf/api/serialization,{toxinidir}/tuf/client_rework
+    pylint {toxinidir}/tuf --ignore={toxinidir}/tuf/api,{toxinidir}/tuf/api/serialization,{toxinidir}/tuf/ngclient
 
     bandit -r {toxinidir}/tuf
