+++ tuf/ngclient/download.py
