+++ tuf/ngclient/fetcher.py
