+++ tuf/ngclient/metadata_bundle.py
